requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
boto3==1.2.2
botocore==1.3.9
celery==3.1.19
cycler==0.9.0
Django==1.8.5
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.2.4
django-rest-auth
docutils==0.12
futures==2.2.0
jmespath==0.9.0
jobtastic==0.3.1
kombu==3.0.29
-e git+https://github.com/projectjamjar/lilo.git@master#egg=lilo-master
Markdown==2.6.2
matplotlib==1.5.1
mysqlclient==1.3.6
numpy==1.10.4
psutil==3.3.0
PyAudio==0.2.9
pydub==0.16.3
pyparsing==2.0.7
python-dateutil==2.4.2
pytz==2015.7
redis==2.10.5
requests==2.8.1
scipy==0.16.1
six==1.10.0
wheel==0.24.0
django-extensions==1.5.9
ipdb==0.8.1
